requirements.txt
Flask
flask-cors
tensorflow
tensorflow-hub
faiss-cpu